+++ idea/testData/refactoring/nameSuggestionProvider/localVarAsCallArgument.kt
@@ -1,0 +1,9 @@
+// SUGGESTED_NAMES: baz, paramBar, s
+fun foo(paramBar: String) {
+
+}
+
+fun test() {
+    val <caret>baz = ""
+    foo(baz)
+}